+++ src/site/twitch.tv/modules/emote-menu/EmoteMenuTab.vue
@@ -134,7 +134,7 @@ const filterSets = debounceFn(() => {
 	// Sort emote sets
 	ary.sort(sortFn);
 	sortedSets.value = ary;
-}, 25);
+}, 500);
 
 // Watch for changes to the emote sets and perform sorting operations
 watch(() => [ctx.filter, sets, cosmetics.emoteSets], filterSets, {